requirements.txt
discord-py-interactions
python-dotenv
spotipy
pytz
pandas
requests
python-dotenv
prettytable
Pillow
asynssh
pydub